+++ Source/DataTransfer.Legacy.MassImport/Core/TelemetryConstants.cs
@@ -9,5 +9,7 @@ namespace DataTransfer.Legacy.MassImport.Core
 		public const string RelativityTelemetrySection = "Relativity.Telemetry";
 		public const string ReleyeUriTracesSettingName = "ReleyeUriTraces";
 		public const string ReleyeTokenSettingName = "ReleyeToken";
+		public const string InstanceIdentifierSettingName = "InstanceIdentifier";
+		public const string RelativityCoreSection = "Relativity.Core";
 	}
 }