+++ Source/DataTransfer.Legacy.Services/ExternalServices/RetryPolicies/IKeplerRetryPolicyFactory.cs
@@ -1,0 +1,10 @@
+﻿using Polly;
+
+namespace Relativity.DataTransfer.Legacy.Services.ExternalServices.RetryPolicies
+{
+	internal interface IKeplerRetryPolicyFactory
+	{
+		IAsyncPolicy CreateRetryPolicy();
+		IAsyncPolicy<T> CreateRetryPolicy<T>();
+	}
+}