+++ Source/DataTransfer.Legacy.Services/Helpers/IRedactedNativesValidator.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Relativity.DataTransfer.Legacy.Services.Helpers
+{
+	public interface IRedactedNativesValidator
+	{
+		Task VerifyThatProductionDoesNotHaveRedactedNativesEnabledAsync(int workspaceID, int productionArtifactID, CancellationToken cancellationToken);
+	}
+}